# swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Txing AI API文档",
        "title": "Txing AI API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/user/list": {
            "get": {
                "description": "获取用户列表，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式(asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态(0:启用, 1:禁用)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/admin/user/status/{id}": {
            "put": {
                "description": "启用或禁用指定用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "切换用户状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/api/user/login": {
            "post": {
                "description": "用户登录并返回token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.LoginVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/user/password": {
            "put": {
                "description": "修改当前登录用户的密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "密码信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/user/profile": {
            "put": {
                "description": "更新当前登录用户的个人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新个人信息",
                "parameters": [
                    {
                        "description": "个人信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateProfileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.UserVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/refresh": {
            "post": {
                "description": "使用刷新令牌获取新的访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "刷新访问令牌",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 刷新令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.TokenPair"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "description": "新用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vo.UserVO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/reset-password": {
            "post": {
                "description": "通过邮箱重置密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "重置密码",
                "parameters": [
                    {
                        "description": "重置密码信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResetPasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.LoginReq": {
            "type": "object",
            "required": [
                "captcha",
                "captchaId",
                "password",
                "username"
            ],
            "properties": {
                "captcha": {
                    "description": "验证码 Captcha",
                    "type": "string",
                    "example": "1234"
                },
                "captchaId": {
                    "description": "验证码ID Captcha ID",
                    "type": "string",
                    "example": "abc123"
                },
                "password": {
                    "description": "密码 Password",
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "description": "用户名 Username",
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "dto.RegisterReq": {
            "type": "object",
            "required": [
                "captcha",
                "captchaId",
                "email",
                "password",
                "username"
            ],
            "properties": {
                "captcha": {
                    "description": "验证码 Captcha",
                    "type": "string",
                    "example": "1234"
                },
                "captchaId": {
                    "description": "验证码ID Captcha ID",
                    "type": "string",
                    "example": "abc123"
                },
                "email": {
                    "description": "邮箱 Email",
                    "type": "string",
                    "example": "john@example.com"
                },
                "password": {
                    "description": "密码 Password",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 6,
                    "example": "password123"
                },
                "phone": {
                    "description": "手机号 Phone number",
                    "type": "string",
                    "example": "13800138000"
                },
                "username": {
                    "description": "用户名 Username",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2,
                    "example": "johndoe"
                }
            }
        },
        "dto.ResetPasswordReq": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "邮箱 Email",
                    "type": "string",
                    "example": "john@example.com"
                }
            }
        },
...
        "dto.UpdatePasswordReq": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码 New password",
                    "type": "string",
                    "example": "newpass123"
                },
                "old_password": {
                    "description": "原密码 Old password",
                    "type": "string",
                    "example": "oldpass123"
                }
            }
        },
...
        "dto.UpdateProfileReq": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄 Age",
                    "type": "integer",
                    "example": 18
                },
                "avatar": {
                    "description": "头像 Avatar URL",
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "email": {
                    "description": "邮箱 Email",
                    "type": "string",
                    "example": "john@example.com"
                },
                "gender": {
                    "description": "性别 Gender: 0-未知 1-男 2-女",
                    "type": "integer",
                    "example": 1
                },
                "phone": {
                    "description": "手机号 Phone number",
                    "type": "string",
                    "example": "13800138000"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "vo.LoginVO": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer",
                    "example": 18
                },
                "avatar": {
                    "description": "头像",
                    "type": "string",
                    "example": "https://example.com/avatar.png"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "example": "a@b.com"
                },
                "gender": {
                    "description": "性别(0:未知 1:男 2:女)",
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string",
                    "example": "13800138000"
                },
                "refreshToken": {
                    "description": "refresh token",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "role": {
                    "description": "角色(0:普通用户 1:超管)",
                    "type": "integer",
                    "example": 0
                },
                "status": {
                    "description": "状态(0:正常 1:禁用)",
                    "type": "integer",
                    "example": 0
                },
                "token": {
                    "description": "access token",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        },
...
        "vo.TokenPair": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "vo.UserVO": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer",
                    "example": 18
                },
                "avatar": {
                    "description": "头像",
                    "type": "string",
                    "example": "https://example.com/avatar.png"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "example": "a@b.com"
                },
                "gender": {
                    "description": "性别(0:未知 1:男 2:女)",
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string",
                    "example": "13800138000"
                },
                "role": {
                    "description": "角色(0:普通用户 1:超管)",
                    "type": "integer",
                    "example": 0
                },
                "status": {
                    "description": "状态(0:正常 1:禁用)",
                    "type": "integer",
                    "example": 0
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        },
...
    }
}